Module rustracing::carrier

Traits for representing carriers that propagate span contexts across process boundaries.

Traits

ExtractFromBinary
This trait allows to extract SpanContext from binary stream.

ExtractFromHttpHeader
This trait allows to extract SpanContext from HTTP header.

ExtractFromTextMap
This trait allows to extract SpanContext from TextMap.

InjectToBinary
This trait allows to inject SpanContext to binary stream.

InjectToHttpHeader
This trait allows to inject SpanContext to HTTP header.

InjectToTextMap
This trait allows to inject SpanContext to TextMap.

IterHttpHeaderFields
This trait allows to iterate over the fields of a HTTP header.

SetHttpHeaderField
This trait allows to insert fields in a HTTP header.

TextMap
This trait represents carriers which support Text Map format.