Trait rustracing::carrier::TextMap

pub trait TextMap {
    fn set(&mut self, key: &str, value: &str);

    fn get(&self, key: &str) -> Option<&str>;
}

This trait represents carriers which support Text Map format.

Text Map is an arbitrary string-to-string map with an unrestricted character set for both keys and values.

Required methods

fn set(&mut self, key: &str, value: &str)

Sets the value of key in the map to value.

fn get(&self, key: &str) -> Option<&str>

Gets the value of `key'.

Loading content...